+++ Frends.SFTP.UploadFiles/Frends.SFTP.UploadFiles/Definitions/Util.cs
@@ -27,7 +27,7 @@ internal static class Util
             pattern = mask.Replace(".", "\\.");
             pattern = pattern.Replace("*", ".*");
             pattern = pattern.Replace("?", ".+");
-            pattern = String.Concat("^", pattern, "$");
+            pattern = string.Concat("^", pattern, "$");
         }
 
         return Regex.IsMatch(filename, pattern, RegexOptions.IgnoreCase);
@@ -38,11 +38,14 @@ internal static class Util
         return fingerprint.Split(':').Select(s => Convert.ToByte(s, 16)).ToArray();
     }
 
-    internal static string ToHex(byte[] bytes)
+    internal static string ToHex(byte[] bytes, CancellationToken cancellationToken)
     {
         StringBuilder result = new StringBuilder(bytes.Length * 2);
         for (int i = 0; i < bytes.Length; i++)
+        {
+            cancellationToken.ThrowIfCancellationRequested();
             result.Append(bytes[i].ToString("x2"));
+        }
         return result.ToString();
     }
 
@@ -60,19 +63,9 @@ internal static class Util
         catch { return false; }
     }
 
-    internal static byte[] ConvertHexStringToHex(string hex)
-    {
-        var arr = new byte[hex.Length / 2];
-        for (var i = 0; i < arr.Length; i++)
-        {
-            arr[i] = Convert.ToByte(hex.Substring(i * 2, 2), 16);
-        }
-        return arr;
-    }
-
     internal static bool IsMD5(string input)
     {
-        if (String.IsNullOrEmpty(input))
+        if (string.IsNullOrEmpty(input))
         {
             return false;
         }
@@ -82,7 +75,7 @@ internal static class Util
 
     internal static bool IsSha256(string input)
     {
-        if (String.IsNullOrEmpty(input))
+        if (string.IsNullOrEmpty(input))
         {
             return false;
         }